# artifacts/api-server/src/routes/menu.ts
import { Router, type IRouter } from "express";
import { eq, sql, inArray } from "drizzle-orm";
import { proxyImageUrl, prewarmImageCache } from "./image-proxy";
import { db, menuCategoriesTable, menuItemsTable, modifiersTable, orderItemsTable } from "@workspace/db";
import { pushSoldOutItemToCloud, pushSoldOutModifierOptionToCloud } from "../lib/online-orders-sync";
import { isStaffAuthenticated } from "./auth";
import {
  CreateMenuCategoryBody,
  UpdateMenuCategoryParams,
  UpdateMenuCategoryBody,
  DeleteMenuCategoryParams,
  ListMenuItemsQueryParams,
  CreateMenuItemBody,
  GetMenuItemParams,
  UpdateMenuItemParams,
  UpdateMenuItemBody,
  DeleteMenuItemParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

// ---- Categories ----

router.get("/menu/categories", async (_req, res): Promise<void> => {
  // Return ALL categories — this endpoint is ground truth and is consumed by
  // the admin menu page (which MUST see empty categories so newly created ones
  // are visible — otherwise the admin can't add the first item to them and
  // ends up creating duplicates), the POS, and the KDS as well as the customer
  // home page. Customer-facing filtering of empty categories happens
  // client-side in home.tsx using the items list it already fetches.
  // (Previously this used an INNER JOIN on menu_items WHERE available=true,
  // which silently hid every empty category from EVERYONE — that's the bug
  // that produced 4 duplicate "Misc" rows in prod on 2026-05-12.)
  const categories = await db
    .select()
    .from(menuCategoriesTable)
    .orderBy(menuCategoriesTable.sortOrder);
  res.set("Cache-Control", "public, max-age=30, stale-while-revalidate=120");
  res.json(categories);
});

router.post("/menu/categories", async (req, res): Promise<void> => {
  const parsed = CreateMenuCategoryBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [category] = await db
    .insert(menuCategoriesTable)
    .values({
      name: parsed.data.name,
      description: parsed.data.description ?? null,
      icon: parsed.data.icon ?? null,
      sortOrder: parsed.data.sortOrder ?? 0,
    })
    .returning();
  res.status(201).json(category);
});

router.patch("/menu/categories/:id", async (req, res): Promise<void> => {
  const params = UpdateMenuCategoryParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = UpdateMenuCategoryBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const updates: Record<string, unknown> = {};
  if (parsed.data.name !== undefined) updates.name = parsed.data.name;
  if (parsed.data.description !== undefined) updates.description = parsed.data.description;
  if (parsed.data.sortOrder !== undefined) updates.sortOrder = parsed.data.sortOrder;
  if (parsed.data.sendToKds !== undefined) updates.sendToKds = parsed.data.sendToKds;
  if (parsed.data.icon !== undefined) updates.icon = parsed.data.icon ?? null;

  const [updated] = await db
    .update(menuCategoriesTable)
    .set(updates)
    .where(eq(menuCategoriesTable.id, params.data.id))
    .returning();
  if (!updated) {
    res.status(404).json({ error: "Category not found" });
    return;
  }
  res.json(updated);
});

router.delete("/menu/categories/:id", async (req, res): Promise<void> => {
  const params = DeleteMenuCategoryParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [deleted] = await db
    .delete(menuCategoriesTable)
    .where(eq(menuCategoriesTable.id, params.data.id))
    .returning();
  if (!deleted) {
    res.status(404).json({ error: "Category not found" });
    return;
  }
  res.sendStatus(204);
});

// ---- Menu Items ----

router.get("/menu/items", async (req, res): Promise<void> => {
  const queryParsed = ListMenuItemsQueryParams.safeParse(req.query);
  if (!queryParsed.success) {
    res.status(400).json({ error: queryParsed.error.message });
    return;
  }
  let query = db.select().from(menuItemsTable).orderBy(menuItemsTable.sortOrder, menuItemsTable.id).$dynamic();
  if (queryParsed.data.categoryId !== undefined) {
    query = query.where(eq(menuItemsTable.categoryId, queryParsed.data.categoryId));
  }
  if (queryParsed.data.available !== undefined) {
    query = query.where(eq(menuItemsTable.available, queryParsed.data.available));
  }
  const items = await query;
  // Hide open-price items (e.g. "Misc") from anonymous callers. They're a POS-only
  // feature with a $0 placeholder price; exposing them publicly enables a low-effort
  // exploit where an attacker reads the item id and POSTs an order with a
  // self-supplied priceOverride. Staff sessions still see them so admin/POS work.
  const staff = isStaffAuthenticated(req);
  // hiddenOnline items are removed from the customer menu entirely (unlike
  // available=false, which shows them greyed out as sold-out). Staff/POS still see them.
  const visible = staff ? items : items.filter((i) => !i.openPrice && !i.hiddenOnline);
  // Pre-warm image cache so all Loyverse images are ready before the browser asks
  prewarmImageCache(visible.flatMap(i => [i.imageUrl, i.posImageUrl]));
  const result = visible.map((item) => ({
    ...item,
    price: parseFloat(item.price as unknown as string),
    imageUrl: proxyImageUrl(item.imageUrl),
    posImageUrl: proxyImageUrl(item.posImageUrl),
  }));
  // No public cache when the response varies by auth — otherwise the proxy could
  // hand a staff-cached body (with open-price items) to an anonymous client.
  // Vary tells well-behaved caches to key on the auth identity too.
  res.set("Vary", "Cookie, Authorization");
  res.set("Cache-Control", staff ? "private, max-age=0, no-store" : "public, max-age=30, stale-while-revalidate=120");
  res.json(result);
});

// ---- Top sellers (based on real order data) ----
// Returns up to `limit` available items ordered by total units sold.
// Falls back gracefully to an empty array if there are no sales yet.
router.get("/menu/popular", async (req, res): Promise<void> => {
  const limit = Math.min(parseInt(String(req.query.limit ?? "5")), 20);
  // Optional rolling window filter via ?days=30
  const days = req.query.days ? parseInt(String(req.query.days)) : null;

  const rows = await db.execute(sql`
    SELECT
      mi.id,
      mi.name,
      mi.description,
      mi.price,
      mi.image_url      AS "imageUrl",
      mi.pos_image_url  AS "posImageUrl",
      mi.available,
      mi.popular,
      mi.spicy,
      mi.vegetarian,
      mi.sort_order     AS "sortOrder",
      mi.category_id    AS "categoryId",
      SUM(oi.quantity)::int AS "totalSold"
    FROM order_items oi
    JOIN orders o ON o.id = oi.order_id
    JOIN menu_items mi ON mi.id = oi.menu_item_id
    WHERE mi.available = true
      AND mi.open_price = false
      AND mi.hidden_online = false
      AND oi.menu_item_id IS NOT NULL
      ${days ? sql`AND o.created_at >= NOW() - (${days} || ' days')::interval` : sql``}
    GROUP BY mi.id
    ORDER BY "totalSold" DESC
    LIMIT ${limit}
  `);

  res.set("Cache-Control", "public, max-age=30, stale-while-revalidate=120");
  res.json(
    rows.rows.map((r) => ({
      ...r,
      price: parseFloat(r.price as string),
      imageUrl: proxyImageUrl(r.imageUrl as string | null),
      posImageUrl: proxyImageUrl(r.posImageUrl as string | null),
    }))
  );
});

router.post("/menu/items", async (req, res): Promise<void> => {
  const parsed = CreateMenuItemBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const openPrice = parsed.data.openPrice ?? false;
  // openPrice items don't carry a fixed price — the cashier sets it at the POS.
  // Coerce missing/zero price safely so the NOT NULL constraint stays satisfied.
  const priceValue = openPrice ? 0 : (parsed.data.price ?? 0);
  if (!openPrice && (parsed.data.price === undefined || parsed.data.price === null)) {
    res.status(400).json({ error: "price is required unless openPrice is true" });
    return;
  }
  const [item] = await db
    .insert(menuItemsTable)
    .values({
      categoryId: parsed.data.categoryId,
      name: parsed.data.name,
      description: parsed.data.description ?? null,
      price: String(priceValue),
      imageUrl: parsed.data.imageUrl ?? null,
      posImageUrl: parsed.data.posImageUrl ?? null,
      available: parsed.data.available ?? true,
      popular: parsed.data.popular ?? false,
      spicy: parsed.data.spicy ?? false,
      vegetarian: parsed.data.vegetarian ?? false,
      openPrice,
      hiddenOnline: parsed.data.hiddenOnline ?? false,
    })
    .returning();
  res.status(201).json({ ...item, price: parseFloat(item.price as unknown as string) });
});

router.get("/menu/items/:id", async (req, res): Promise<void> => {
  const params = GetMenuItemParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [item] = await db
    .select()
    .from(menuItemsTable)
    .where(eq(menuItemsTable.id, params.data.id));
  if (!item) {
    res.status(404).json({ error: "Item not found" });
    return;
  }
  // Open-price and hidden-online items are not public — return 404 to anonymous callers
  // so this endpoint can't be used to fish for the item id (paired with the list filter).
  if ((item.openPrice || item.hiddenOnline) && !isStaffAuthenticated(req)) {
    res.status(404).json({ error: "Item not found" });
    return;
  }
  res.set("Vary", "Cookie, Authorization");
  res.set("Cache-Control", "private, max-age=0, no-store");
  res.json({ ...item, price: parseFloat(item.price as unknown as string), imageUrl: proxyImageUrl(item.imageUrl), posImageUrl: proxyImageUrl(item.posImageUrl) });
});

router.get("/menu/items/:id/modifiers", async (req, res): Promise<void> => {
  const id = parseInt(req.params.id);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }

  const [item] = await db.select().from(menuItemsTable).where(eq(menuItemsTable.id, id));
  if (!item) { res.status(404).json({ error: "Item not found" }); return; }
  // Don't leak open-price/hidden item existence via the modifiers endpoint either.
  if ((item.openPrice || item.hiddenOnline) && !isStaffAuthenticated(req)) {
    res.status(404).json({ error: "Item not found" });
    return;
  }

  const modifierIds = item.loyverseModifierIds ?? [];
  if (modifierIds.length === 0) { res.json([]); return; }

  const mods = await db
    .select()
    .from(modifiersTable)
    .where(inArray(modifiersTable.loyverseId, modifierIds));

  // Return sorted by global sortOrder so POS respects the admin-set order
  // Strip out any options that are currently 86'd (unavailableOptionIds)
  const ordered = mods
    .filter((m) => modifierIds.includes(m.loyverseId))
    .sort((a, b) => a.sortOrder - b.sortOrder || a.name.localeCompare(b.name))
    .map((m) => ({
      ...m,
      options: (m.options as import("@workspace/db").ModifierOption[]).filter(
        (o) => !(m.unavailableOptionIds ?? []).includes(o.id)
      ),
    }));

  res.json(ordered);
});

// ---- 86 List (Sold Out) ----

// GET /api/menu/soldout — full sold-out snapshot for the POS panel
router.get("/menu/soldout", async (_req, res): Promise<void> => {
  const [items, modifiers] = await Promise.all([
    db
      .select({
        id: menuItemsTable.id,
        name: menuItemsTable.name,
        categoryId: menuItemsTable.categoryId,
        available: menuItemsTable.available,
      })
      .from(menuItemsTable)
      .orderBy(menuItemsTable.categoryId, menuItemsTable.sortOrder),
    db
      .select({
        id: modifiersTable.id,
        name: modifiersTable.name,
        options: modifiersTable.options,
        unavailableOptionIds: modifiersTable.unavailableOptionIds,
      })
      .from(modifiersTable)
      .orderBy(modifiersTable.sortOrder, modifiersTable.name),
  ]);
  const categories = await db
    .select({ id: menuCategoriesTable.id, name: menuCategoriesTable.name })
    .from(menuCategoriesTable)
    .orderBy(menuCategoriesTable.sortOrder);
  res.json({ items, modifiers, categories });
});

// POST /api/menu/soldout/item/:id — toggle a menu item's available flag
router.post("/menu/soldout/item/:id", async (req, res): Promise<void> => {
  const id = parseInt(req.params.id);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }
  const { available } = req.body as { available?: boolean };
  if (typeof available !== "boolean") { res.status(400).json({ error: "available (bool) required" }); return; }
  const [updated] = await db
    .update(menuItemsTable)
    .set({ available })
    .where(eq(menuItemsTable.id, id))
    .returning({ id: menuItemsTable.id, available: menuItemsTable.available });
  if (!updated) { res.status(404).json({ error: "Item not found" }); return; }
  pushSoldOutItemToCloud(id, available);
  res.json(updated);
});

// POST /api/menu/soldout/modifier-option — toggle a modifier option's availability
// Body: { modifierId: number, optionId: string, available: boolean }
router.post("/menu/soldout/modifier-option", async (req, res): Promise<void> => {
  const { modifierId, optionId, available } = req.body as { modifierId?: unknown; optionId?: unknown; available?: unknown };
  const mid = Number(modifierId);
  if (isNaN(mid) || typeof optionId !== "string" || !optionId || typeof available !== "boolean") {
    res.status(400).json({ error: "modifierId (number), optionId (string), available (bool) required" });
    return;
  }
  const [modifier] = await db.select().from(modifiersTable).where(eq(modifiersTable.id, mid));
  if (!modifier) { res.status(404).json({ error: "Modifier not found" }); return; }

  let ids: string[] = Array.isArray(modifier.unavailableOptionIds) ? [...modifier.unavailableOptionIds] : [];
  if (!available && !ids.includes(optionId)) {
    ids.push(optionId);
  } else if (available) {
    ids = ids.filter((i) => i !== optionId);
  }

  const [updated] = await db
    .update(modifiersTable)
    .set({ unavailableOptionIds: ids })
    .where(eq(modifiersTable.id, mid))
    .returning({ id: modifiersTable.id, unavailableOptionIds: modifiersTable.unavailableOptionIds });
  pushSoldOutModifierOptionToCloud(mid, optionId, available);
  res.json(updated);
});

router.patch("/menu/items/:id", async (req, res): Promise<void> => {
  const params = UpdateMenuItemParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = UpdateMenuItemBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const updates: Record<string, unknown> = {};
  if (parsed.data.categoryId !== undefined) updates.categoryId = parsed.data.categoryId;
  if (parsed.data.name !== undefined) updates.name = parsed.data.name;
  if (parsed.data.description !== undefined) updates.description = parsed.data.description;
  if (parsed.data.price !== undefined) updates.price = String(parsed.data.price);
  if (parsed.data.imageUrl !== undefined) updates.imageUrl = parsed.data.imageUrl;
  if (parsed.data.posImageUrl !== undefined) updates.posImageUrl = parsed.data.posImageUrl;
  if (parsed.data.available !== undefined) updates.available = parsed.data.available;
  if (parsed.data.popular !== undefined) updates.popular = parsed.data.popular;
  if (parsed.data.spicy !== undefined) updates.spicy = parsed.data.spicy;
  if (parsed.data.vegetarian !== undefined) updates.vegetarian = parsed.data.vegetarian;
  if (parsed.data.hiddenOnline !== undefined) updates.hiddenOnline = parsed.data.hiddenOnline;
  if (parsed.data.openPrice !== undefined) {
    updates.openPrice = parsed.data.openPrice;
    // Open-price items don't carry a fixed price — normalize to 0 so reports/exports
    // stay sane and the storefront price ($0) is never shown (those items are also
    // hidden from the customer menu, see home.tsx).
    if (parsed.data.openPrice === true) updates.price = "0";
    // The "price > 0 required" rule only applies when the caller is actually
    // transitioning open-price OFF (true → false). If the item was already
    // openPrice=false, accept the save as-is — admins routinely re-save existing
    // $0 / placeholder rows (e.g. legacy Misc items) without touching price.
    // Without this guard the form is permanently un-savable for any existing
    // $0 fixed-price row, since the admin UI always sends `openPrice` in the body.
    if (parsed.data.openPrice === false) {
      const p = parsed.data.price;
      const newPositivePriceProvided = typeof p === "number" && p > 0;
      if (!newPositivePriceProvided) {
        const [existing] = await db
          .select({ openPrice: menuItemsTable.openPrice })
          .from(menuItemsTable)
          .where(eq(menuItemsTable.id, params.data.id))
          .limit(1);
        if (!existing) {
          res.status(404).json({ error: "Item not found" });
          return;
        }
        if (existing.openPrice === true) {
          res.status(400).json({ error: "price (> 0) is required when disabling openPrice" });
          return;
        }
      }
    }
  }
  const body = req.body as Record<string, unknown>;
  if (body["sortOrder"] !== undefined) updates.sortOrder = body["sortOrder"];
  if (body["loyverseModifierIds"] !== undefined) updates.loyverseModifierIds = body["loyverseModifierIds"] ?? null;

  const [item] = await db
    .update(menuItemsTable)
    .set(updates)
    .where(eq(menuItemsTable.id, params.data.id))
    .returning();
  if (!item) {
    res.status(404).json({ error: "Item not found" });
    return;
  }
  res.json({ ...item, price: parseFloat(item.price as unknown as string) });
});

// ---- Category reorder ----
// Accepts { ids: number[] } — the full ordered list of category IDs.
router.post("/menu/categories/reorder", async (req, res): Promise<void> => {
  const { ids } = req.body as { ids?: unknown };
  if (!Array.isArray(ids) || ids.some((v) => typeof v !== "number")) {
    res.status(400).json({ error: "ids must be an array of numbers" });
    return;
  }
  const typedIds = ids as number[];
  await Promise.all(
    typedIds.map((id, idx) =>
      db.update(menuCategoriesTable).set({ sortOrder: idx * 10 }).where(eq(menuCategoriesTable.id, id))
    )
  );
  res.json({ ok: true });
});

// ---- Bulk reorder ----
// Accepts { ids: number[] } — the full ordered list of item IDs.
// Assigns sortOrder = index * 10, persisting the drag-drop sequence.
router.post("/menu/items/reorder", async (req, res): Promise<void> => {
  const { ids } = req.body as { ids?: unknown };
  if (!Array.isArray(ids) || ids.some((v) => typeof v !== "number")) {
    res.status(400).json({ error: "ids must be an array of numbers" });
    return;
  }
  const typedIds = ids as number[];
  await Promise.all(
    typedIds.map((id, idx) =>
      db.update(menuItemsTable).set({ sortOrder: idx * 10 }).where(eq(menuItemsTable.id, id))
    )
  );
  res.json({ ok: true });
});

router.delete("/menu/items/:id", async (req, res): Promise<void> => {
  const params = DeleteMenuItemParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [deleted] = await db
    .delete(menuItemsTable)
    .where(eq(menuItemsTable.id, params.data.id))
    .returning();
  if (!deleted) {
    res.status(404).json({ error: "Item not found" });
    return;
  }
  res.sendStatus(204);
});

// ---- Modifiers CRUD ----

router.get("/menu/modifiers", async (_req, res): Promise<void> => {
  const modifiers = await db
    .select()
    .from(modifiersTable)
    .orderBy(modifiersTable.sortOrder, modifiersTable.name);
  res.json(modifiers);
});

// ---- Bulk reorder modifiers ----
router.post("/menu/modifiers/reorder", async (req, res): Promise<void> => {
  const { ids } = req.body as { ids?: unknown };
  if (!Array.isArray(ids) || ids.some((v) => typeof v !== "number")) {
    res.status(400).json({ error: "ids must be an array of numbers" });
    return;
  }
  const typedIds = ids as number[];
  await Promise.all(
    typedIds.map((id, idx) =>
      db.update(modifiersTable).set({ sortOrder: idx * 10 }).where(eq(modifiersTable.id, id))
    )
  );
  res.json({ ok: true });
});

router.post("/menu/modifiers", async (req, res): Promise<void> => {
  const { name, options, required, minSelections, maxSelections } = req.body as {
    name?: string; options?: unknown[]; required?: boolean; minSelections?: number; maxSelections?: number | null;
  };
  if (!name || typeof name !== "string" || !name.trim()) {
    res.status(400).json({ error: "name is required" });
    return;
  }
  const loyverseId = `manual_${crypto.randomUUID()}`;
  const [modifier] = await db
    .insert(modifiersTable)
    .values({
      loyverseId,
      name: name.trim(),
      options: (options ?? []) as import("@workspace/db").ModifierOption[],
      required: required ?? false,
      minSelections: minSelections ?? 0,
      maxSelections: maxSelections ?? null,
    })
    .returning();
  res.status(201).json(modifier);
});

router.patch("/menu/modifiers/:id", async (req, res): Promise<void> => {
  const id = parseInt(req.params.id);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }
  const { name, options, required, minSelections, maxSelections } = req.body as {
    name?: string; options?: unknown[]; required?: boolean; minSelections?: number; maxSelections?: number | null;
  };
  const updates: Record<string, unknown> = {};
  if (name !== undefined) updates.name = String(name).trim();
  if (options !== undefined) updates.options = options;
  if (required !== undefined) updates.required = required;
  if (minSelections !== undefined) updates.minSelections = minSelections;
  if (maxSelections !== undefined) updates.maxSelections = maxSelections === null ? null : Number(maxSelections);
  if (!Object.keys(updates).length) { res.status(400).json({ error: "Nothing to update" }); return; }
  const [modifier] = await db
    .update(modifiersTable)
    .set(updates)
    .where(eq(modifiersTable.id, id))
    .returning();
  if (!modifier) { res.status(404).json({ error: "Modifier not found" }); return; }
  res.json(modifier);
});

router.delete("/menu/modifiers/:id", async (req, res): Promise<void> => {
  const id = parseInt(req.params.id);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }
  const [deleted] = await db
    .delete(modifiersTable)
    .where(eq(modifiersTable.id, id))
    .returning();
  if (!deleted) { res.status(404).json({ error: "Modifier not found" }); return; }
  res.sendStatus(204);
});

export default router;