# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Island Tacos Online Ordering API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"
    post:
      operationId: createMenuCategory
      tags: [menu]
      summary: Create a menu category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
  /menu/categories/{id}:
    patch:
      operationId: updateMenuCategory
      tags: [menu]
      summary: Update a menu category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMenuCategory
      tags: [menu]
      summary: Delete a menu category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items (optionally filtered by category)
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MenuCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
        sendToKds:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sortOrder
        - sendToKds
        - createdAt
    CreateMenuCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
        sendToKds:
          type: boolean
      required:
        - name
    UpdateMenuCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
        sendToKds:
          type: boolean
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        posImageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        popular:
          type: boolean
        spicy:
          type: boolean
        vegetarian:
          type: boolean
        openPrice:
          type: boolean
        hiddenOnline:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - categoryId
        - name
        - price
        - available
        - popular
        - spicy
        - vegetarian
        - openPrice
        - hiddenOnline
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        posImageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        popular:
          type: boolean
        spicy:
          type: boolean
        vegetarian:
          type: boolean
        openPrice:
          type: boolean
        hiddenOnline:
          type: boolean
      required:
        - categoryId
        - name
    UpdateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        posImageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        popular:
          type: boolean
        spicy:
          type: boolean
        vegetarian:
          type: boolean
        openPrice:
          type: boolean
        hiddenOnline:
          type: boolean
...