# lib/api-zod/src/generated/types/updateMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Island Tacos Online Ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMenuItemBody {
  categoryId?: number;
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  posImageUrl?: string | null;
  available?: boolean;
  popular?: boolean;
  spicy?: boolean;
  vegetarian?: boolean;
  openPrice?: boolean;
  hiddenOnline?: boolean;
}